BrendaCore/requirements_phase2.txt
# Phase 2 Requirements - Cartesia Voice Integration

# Core dependencies from Phase 1
pyyaml>=6.0
aiohttp>=3.8.0

# Voice and Audio Processing
# Note: Cartesia SDK will be required when available
# cartesia-ai>=1.0.0  # Placeholder - use when SDK is released

# Async and Networking
asyncio>=3.4.3
aiofiles>=23.0.0
websockets>=11.0

# Security and Authentication
cryptography>=41.0.0
pyjwt>=2.8.0

# Configuration and Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-aiohttp>=1.0.0

# Monitoring and Logging
prometheus-client>=0.18.0

# Optional: For enhanced voice features
# numpy>=1.24.0  # For audio processing
# scipy>=1.11.0  # For signal processing